chore: refactor into more components
