+++ package.json
@@ -38,7 +38,8 @@
     "react-native-svg": "13.9.0",
     "react-native-web": "~0.19.6",
     "viem": "^1.19.0",
-    "wagmi": "^1.4.6"
+    "wagmi": "^1.4.6",
+    "expo-secure-store": "~12.3.1"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
