+++ src/lib/lens-sdk/index.ts
@@ -1,13 +1,16 @@
 import { LensConfig, development } from '@lens-protocol/react-native';
 import { bindings as wagmiBindings } from '@lens-protocol/wagmi';
-import AsyncStorage from '@react-native-async-storage/async-storage';
+import * as SecureStore from 'expo-secure-store';
 
 const lensConfig: LensConfig = {
   bindings: wagmiBindings(),
   // Use `development` or `production` depending on your need
   environment: development,
-  // TODO: Check if security is an issue - IDEA: SecureStorage
-  storage: AsyncStorage,
+  storage: {
+    getItem: SecureStore.getItemAsync,
+    setItem: SecureStore.setItemAsync,
+    removeItem: SecureStore.deleteItemAsync,
+  },
 };
 
 export * from '@lens-protocol/react-native';
