+++ src/lib/theme.ts
@@ -19,6 +19,7 @@ const theme = {
     disabled: 'rgb(66,64,64)',
     danger: 'rgb(255, 69, 58)',
     overlay: 'rgba(0, 0, 0, 0.7)',
+    lightWhite: 'rgba(255, 255, 255, 0.1)',
   },
 };
